SensorDto's ready

+++ src/main/java/skni/kamilG/skin_sensors_api/Model/Sensor/SensorData.java
@@ -24,12 +24,12 @@ public class SensorData {
 
   private LocalDateTime timestamp;
 
-  private short temperature;
+  private Short temperature;
 
-  private long humidity;
+  private Integer humidity;
 
-  private long pressure;
+  private Integer pressure;
 
   @Column(name = "gas_resistance", nullable = false)
-  private long gasResistance;
+  private int gasResistance;
 }
